+++ machine_learning/src/face_detect_model/DetectFaceAndClip/detectFaceUtil.py
@@ -1,5 +1,4 @@
 import cv2
-import os
 
 
 def load_cascade(cascade_path):
