+++ machine_learning/src/face_detect_model/pred.py
@@ -1,0 +1,4 @@
+def main():
+    gray = cv2.cvtColor(captureBuffer, cv2.COLOR_BGR2GRAY)
+    pred_child_ids = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+    return pred_child_ids
