+++ machine_learning/src/face_detect_model/util.py
@@ -3,7 +3,6 @@ import re
 
 import random
 import torch
-import google.cloud.storage as gcs
 from google.cloud.storage import Bucket
 import os
 import numpy as np
@@ -23,39 +22,6 @@ def set_seed(seed):
     torch.manual_seed(seed)
     torch.cuda.manual_seed(seed)
     torch.cuda.manual_seed_all(seed)
-
-
-def init_client():
-    # NOTE: gcloud auth application-default loginにて事前に認証
-    PROJECT_ID = os.environ.get("PROJECT_ID")
-
-    client = gcs.Client(PROJECT_ID)
-    if client is None:
-        logger.error("Failed to initialize client.")
-        exit(1)
-    else:
-        return client
-
-
-def get_bucket(client: gcs.Client, bucket_name: str):
-    bucket = client.bucket(bucket_name)
-
-    if bucket.exists():
-        return bucket
-    else:
-        raise ValueError(f"Failed to {bucket_name} does not exist.")
-
-
-def get_blobs(bucket: Bucket, blob_name: str):
-    # blobsの中身に対するエラーハンドリング
-    try:
-        blobs = list(bucket.list_blobs(prefix=blob_name))
-        if len(blobs) == 0:  # 最初の要素がない場合、イテレータは空
-            raise ValueError(f"No blobs found with prefix '{blob_name}' in the bucket.")
-        else:
-            return blobs
-    except Exception as e:
-        raise ValueError(f"Failed to get blobs from '{blob_name}' due to an error: {e}")
 
 
 def get_child_id(blob_name: str):
